avahidbus: add call to create new service browser on unicast refresh
Thought this was effectively being done by removing/adding watchers,
but turns out this wasn't the case. Unfortunately without a way to
free resources in Avahi via DBus you eventually consume them all.

Probably need to re-think this class for unicast DNS purposes, closing
the DBus connection in the unicast case only when the 'AllForNow' signal
is received via DBus, and providing a mechanism to prompt a new connection
and browse.