aggregator: linting fix which could affect behaviour
This line is intended to force a re-discovery of a Reg API if
we receive a 4XX error from it during initial (re-)registration.
Assuming our Node schema validates this will only occur due to an
implementation error in the Reg API and is intended to prevent
use getting stuck contacting that instance when others may work
properly.

The line has clearly never done its job, but it doesn't appear
that enabling it will do any harm as in the worst case it will
cause the same Reg API to be found again via the mDNS bridge on
the next iteration.